add controllers and initializer

+++ src/main/java/com/example/onlinecinema/service/impl/MovieServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.onlinecinema.service.impl;
 
 import com.example.onlinecinema.domain.exeption.ResourceNotFoundException;
 import com.example.onlinecinema.domain.movie.Movie;
+import com.example.onlinecinema.domain.session.MovieSession;
 import com.example.onlinecinema.repository.MovieRepository;
 import com.example.onlinecinema.service.interfaces.MovieService;
 import lombok.RequiredArgsConstructor;
@@ -19,18 +20,27 @@ public class MovieServiceImpl implements MovieService {
     @Override
     public Movie create(Movie movie) {
         if (movieRepository.findByTitle(movie.getTitle()).isPresent()) {
-            throw new IllegalStateException("User already exist");
+            throw new IllegalStateException("Movie already exist");
         }
         return movieRepository.save(movie);
     }
 
     @Override
     public Movie update(Movie movie) {
-        return movieRepository.save(movie);
+
+        Movie validatedMovie = movieRepository.findById(movie.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Movie by id not found exception(update movie method)"));
+        validatedMovie.setTitle(movie.getTitle());
+        validatedMovie.setDirector(movie.getDirector());
+        validatedMovie.setPoster(movie.getPoster());
+        validatedMovie.setDurationMinutes(movie.getDurationMinutes());
+        return movieRepository.save(validatedMovie);
     }
 
     @Override
     public void deleteById(Long id) {
+        if(movieRepository.findById(id).isEmpty())
+            throw new IllegalStateException("Movie already delete");
         movieRepository.deleteById(id);
     }
 
@@ -45,4 +55,19 @@ public class MovieServiceImpl implements MovieService {
         return new ArrayList<>(movieRepository.findAll());
     }
 
+    @Override
+    public MovieSession createMovieSession(Long id, MovieSession movieSession) {
+
+        movieSession.createTickets();
+
+        Movie movie = getMovieById(id);
+
+        if(movie.getMovieSessions().isEmpty()) {
+            movie.setMovieSessions(new ArrayList<>());
+        }
+
+        movie.getMovieSessions().add(movieSession);
+        movieRepository.save(movie);
+        return movieSession;
+    }
 }
